Classify agi-gui modules into subpackages
