+++ src/agilab/lib/agi-gui/src/agi_gui/compat/__init__.py
@@ -1,0 +1,7 @@
+"""Compatibility helpers for agi-gui public module aliases."""
+
+from __future__ import annotations
+
+from .module_alias import activate_compat_module, alias_agi_env_module
+
+__all__ = ["activate_compat_module", "alias_agi_env_module"]
